requirements.txt
biopython==1.79
numpy==1.21.5
scipy==1.7.3
